+++ contracts/hackatom/tests/integration.rs
@@ -43,9 +43,9 @@ fn proper_initialization() {
     let beneficiary = HumanAddr(String::from("benefits"));
     let creator = HumanAddr(String::from("creator"));
     let expected_state = State {
-        verifier: deps.api.canonical_address(&verifier).unwrap(),
-        beneficiary: deps.api.canonical_address(&beneficiary).unwrap(),
-        funder: deps.api.canonical_address(&creator).unwrap(),
+        verifier: deps.api.canonical_address(&verifier).unwrap().0,
+        beneficiary: deps.api.canonical_address(&beneficiary).unwrap().0,
+        funder: deps.api.canonical_address(&creator).unwrap().0,
     };
 
     let msg = InitMsg {
@@ -183,7 +183,7 @@ fn handle_release_works() {
     };
     let init_amount = coins(1000, "earth");
     let init_env = mock_env(&deps.api, creator.as_str(), &init_amount);
-    let contract_addr = deps.api.human_address(&init_env.contract.address).unwrap();
+    let (contract_addr, _gas_used) = deps.api.human_address(&init_env.contract.address).unwrap();
     let init_res: InitResponse = init(&mut deps, init_env, init_msg).unwrap();
     assert_eq!(init_res.messages.len(), 0);
 
@@ -230,7 +230,7 @@ fn handle_release_fails_for_wrong_sender() {
     };
     let init_amount = coins(1000, "earth");
     let init_env = mock_env(&deps.api, creator.as_str(), &init_amount);
-    let contract_addr = deps.api.human_address(&init_env.contract.address).unwrap();
+    let (contract_addr, _gas_used) = deps.api.human_address(&init_env.contract.address).unwrap();
     let init_res: InitResponse = init(&mut deps, init_env, init_msg).unwrap();
     assert_eq!(init_res.messages.len(), 0);
 
@@ -263,9 +263,9 @@ fn handle_release_fails_for_wrong_sender() {
     assert_eq!(
         state,
         State {
-            verifier: deps.api.canonical_address(&verifier).unwrap(),
-            beneficiary: deps.api.canonical_address(&beneficiary).unwrap(),
-            funder: deps.api.canonical_address(&creator).unwrap(),
+            verifier: deps.api.canonical_address(&verifier).unwrap().0,
+            beneficiary: deps.api.canonical_address(&beneficiary).unwrap().0,
+            funder: deps.api.canonical_address(&creator).unwrap().0,
         }
     );
 }